Added option to cast some columns to type expected by Tarantool (#24)

+++ docker/mysql/init.d/init.sql
@@ -14,7 +14,7 @@ CREATE TABLE logins
     username  varchar(16)  not null,
     ip        varchar(16)  not null,
     date      int unsigned not null,
-    attempts  int unsigned   default 0,
+    attempts  bigint(20)     default 0,
     longitude float unsigned default 0,
     latitude  float unsigned default 0,
